Fix warnings

+++ src/Lt/Graphics/GL1Render.cpp
@@ -7,7 +7,7 @@ Lt::Graphics::GL1Render::GL1Render(Lt::Core::ErrorHandler& errorHandler, Lt::Gra
 {
 	glMatrixMode(GL_PROJECTION);
 	glLoadIdentity();
-	glOrtho(0.0f, _Window->Size().PosX(), _Window->Size().PosY(), 0.0f, 0.0f, 1.0f);
+	glOrtho(0.0f, (GLdouble)_Window->Size().PosX(), (GLdouble)_Window->Size().PosY(), 0.0f, 0.0f, 1.0f);
 	glMatrixMode(GL_MODELVIEW);
 	glEnable(GL_TEXTURE_2D);
 }
